Store parsed rule actions into a class
Following a fail-first method, the  method will take a modmail conversation, subreddit and its state, and return whether the bot should run actions based on its result.
Within the method, it'll check all the parsed actions, which was sorted by priority, and determine whether the modmail conversation needs to be actioned upon.